Refactor request handling and utility functions
- Moved request handling functions from tcp_server_refactored.cpp to a new header file request_handlers.hpp for better organization and modularity.
- Updated tcp_server_refactored.cpp to include the new request_handlers.hpp and removed redundant utility functions.
- Consolidated utility functions related to JSON response handling and password security into request_handlers.hpp.
- Improved code readability and maintainability by separating concerns and reducing the size of tcp_server_refactored.cpp.
